+++ src/Components/TransactionHandler/Capture/CaptureTransactionHandler.php
@@ -51,10 +51,24 @@ class CaptureTransactionHandler extends AbstractTransactionHandler implements Ca
         $this->updateTransactionData($parameterBag, (float) $parameterBag->get('amount'));
         $this->saveOrderLineItemData($parameterBag->get('orderLines', []), $context);
 
-        $transitionName = StateMachineTransitionActions::ACTION_PAY_PARTIALLY;
+        $customFields = $this->paymentTransaction->getCustomFields();
+        $clearingType = strtolower($customFields[$this->getClearingTypeCustomField()] ?? '');
+
+        // Filter payment methods that do not allow changing transaction status at this point
+        if (!in_array($clearingType, ['vor'], true)) {
+            // Update the transaction status if PAYONE capture request was approved
+            $this->updateTransactionStatus($parameterBag, $context);
+        }
+
+        return $requestResponse;
+    }
+
+    protected function updateTransactionStatus(ParameterBag $parameterBag, Context $context): void
+    {
+        $transitionName = StateMachineTransitionActions::ACTION_PAID_PARTIALLY;
 
         if ($parameterBag->get('complete')) {
-            $transitionName = StateMachineTransitionActions::ACTION_PAY;
+            $transitionName = StateMachineTransitionActions::ACTION_PAID;
         }
 
         $this->transactionStatusService->transitionByName(
@@ -62,8 +76,6 @@ class CaptureTransactionHandler extends AbstractTransactionHandler implements Ca
             $this->paymentTransaction->getOrderTransaction()->getId(),
             $transitionName
         );
-
-        return $requestResponse;
     }
 
     protected function getAmountCustomField(): string
@@ -79,5 +91,10 @@ class CaptureTransactionHandler extends AbstractTransactionHandler implements Ca
     protected function getAllowCustomField(): string
     {
         return CustomFieldInstaller::ALLOW_CAPTURE;
+    }
+
+    protected function getClearingTypeCustomField(): string
+    {
+        return CustomFieldInstaller::CLEARING_TYPE;
     }
 }