+++ src/Installer/ConfigInstaller.php
@@ -5,16 +5,12 @@ declare(strict_types=1);
 namespace PayonePayment\Installer;
 
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\Plugin\Context\ActivateContext;
 use Shopware\Core\Framework\Plugin\Context\DeactivateContext;
 use Shopware\Core\Framework\Plugin\Context\InstallContext;
 use Shopware\Core\Framework\Plugin\Context\UninstallContext;
 use Shopware\Core\Framework\Plugin\Context\UpdateContext;
 use Shopware\Core\System\StateMachine\Aggregation\StateMachineTransition\StateMachineTransitionActions;
-use Shopware\Core\System\StateMachine\Aggregation\StateMachineTransition\StateMachineTransitionEntity;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -35,32 +31,37 @@ class ConfigInstaller implements InstallerInterface
 
         // Default payment status mapping
         'paymentStatusAppointed'      => StateMachineTransitionActions::ACTION_REOPEN,
-        'paymentStatusCapture'        => StateMachineTransitionActions::ACTION_PAY,
-        'paymentStatusPartialCapture' => StateMachineTransitionActions::ACTION_PAY_PARTIALLY,
-        'paymentStatusPaid'           => StateMachineTransitionActions::ACTION_PAY,
-        'paymentStatusUnderpaid'      => StateMachineTransitionActions::ACTION_PAY_PARTIALLY,
+        'paymentStatusCapture'        => StateMachineTransitionActions::ACTION_PAID,
+        'paymentStatusPartialCapture' => StateMachineTransitionActions::ACTION_PAID_PARTIALLY,
+        'paymentStatusPaid'           => StateMachineTransitionActions::ACTION_PAID,
+        'paymentStatusUnderpaid'      => StateMachineTransitionActions::ACTION_PAID_PARTIALLY,
         'paymentStatusCancelation'    => StateMachineTransitionActions::ACTION_CANCEL,
         'paymentStatusRefund'         => StateMachineTransitionActions::ACTION_REFUND,
         'paymentStatusPartialRefund'  => StateMachineTransitionActions::ACTION_REFUND_PARTIALLY,
-        'paymentStatusDebit'          => StateMachineTransitionActions::ACTION_PAY,
+        'paymentStatusDebit'          => StateMachineTransitionActions::ACTION_PAID,
         'paymentStatusReminder'       => StateMachineTransitionActions::ACTION_REMIND,
         'paymentStatusVauthorization' => '',
         'paymentStatusVsettlement'    => '',
         'paymentStatusTransfer'       => StateMachineTransitionActions::ACTION_CANCEL,
-        'paymentStatusInvoice'        => StateMachineTransitionActions::ACTION_PAY,
+        'paymentStatusInvoice'        => StateMachineTransitionActions::ACTION_PAID,
         'paymentStatusFailed'         => StateMachineTransitionActions::ACTION_CANCEL,
+    ];
+
+    private const UPDATE_VALUES = [ // Updated for 6.2
+        'paymentStatusCapture'        => [StateMachineTransitionActions::ACTION_PAY => StateMachineTransitionActions::ACTION_PAID],
+        'paymentStatusPartialCapture' => [StateMachineTransitionActions::ACTION_PAY_PARTIALLY => StateMachineTransitionActions::ACTION_PAID_PARTIALLY],
+        'paymentStatusPaid'           => [StateMachineTransitionActions::ACTION_PAY => StateMachineTransitionActions::ACTION_PAID],
+        'paymentStatusUnderpaid'      => [StateMachineTransitionActions::ACTION_PAY_PARTIALLY => StateMachineTransitionActions::ACTION_PAID_PARTIALLY],
+        'paymentStatusDebit'          => [StateMachineTransitionActions::ACTION_PAY => StateMachineTransitionActions::ACTION_PAID],
+        'paymentStatusInvoice'        => [StateMachineTransitionActions::ACTION_PAY => StateMachineTransitionActions::ACTION_PAID],
     ];
 
     /** @var SystemConfigService */
     private $systemConfigService;
 
-    /** @var EntityRepositoryInterface */
-    private $transitionRepository;
-
     public function __construct(ContainerInterface $container)
     {
-        $this->systemConfigService  = $container->get(SystemConfigService::class);
-        $this->transitionRepository = $container->get('state_machine_transition.repository');
+        $this->systemConfigService = $container->get(SystemConfigService::class);
     }
 
     /**
@@ -124,18 +125,21 @@ class ConfigInstaller implements InstallerInterface
                 continue;
             }
 
-            if (strpos($key, 'paymentStatus')) {
-                $transitionCriteria = new Criteria();
-                $transitionCriteria->addAssociation('state_machine');
-                $transitionCriteria->addFilter(new EqualsFilter('actionName', $value));
-                $transitionCriteria->addFilter(new EqualsFilter('technicalName', 'order_transaction.state'));
-
-                /** @var StateMachineTransitionEntity $searchResult */
-                $searchResult = $this->transitionRepository->search($transitionCriteria, $context)->first();
-                $value        = $searchResult->getId();
-            }
-
             $this->systemConfigService->set($configKey, $value);
+        }
+
+        foreach (self::UPDATE_VALUES as $key => $values) {
+            foreach ($values as $from => $to) {
+                $configKey = $domain . $key;
+
+                $currentValue = $this->systemConfigService->get($configKey);
+
+                if ($currentValue !== null && $currentValue !== $from) {
+                    continue;
+                }
+
+                $this->systemConfigService->set($configKey, $to);
+            }
         }
     }
 }