+++ src/Payone/Request/PayolutionInstallment/AbstractPayolutionInstallmentAuthorizeRequestFactory.php
@@ -48,10 +48,13 @@ abstract class AbstractPayolutionInstallmentAuthorizeRequestFactory extends Abst
             $context
         );
 
+        $referenceNumber = $this->systemRequest->getReferenceNumber($transaction, true);
+
         $this->requests[] = $this->payolutionInstallmentRequest->getRequestParameters(
             $transaction,
             $dataBag,
-            $context
+            $context,
+            $referenceNumber
         );
 
         return $this->createRequest();