+++ src/Payone/Request/SofortBanking/SofortBankingAuthorizeRequest.php
@@ -9,9 +9,12 @@ use Shopware\Core\Framework\Context;
 
 class SofortBankingAuthorizeRequest extends AbstractSofortBankingAuthorizeRequest
 {
-    public function getRequestParameters(PaymentTransaction $transaction, Context $context): array
-    {
-        return array_merge(parent::getRequestParameters($transaction, $context), [
+    public function getRequestParameters(
+        PaymentTransaction $transaction,
+        Context $context,
+        string $referenceNumber
+    ): array {
+        return array_merge(parent::getRequestParameters($transaction, $context, $referenceNumber), [
             'request' => 'authorization',
         ]);
     }