+++ tests/Mock/Repository/EntityRepositoryMock.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace PayonePayment\Test\Mock\Repository;
 
 use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\Entity;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityCollection;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
@@ -17,6 +18,14 @@ use Shopware\Core\Framework\Event\NestedEventCollection;
 
 class EntityRepositoryMock implements EntityRepositoryInterface
 {
+    /** @var null|Entity */
+    private $entity;
+
+    public function __construct(?Entity $entity)
+    {
+        $this->entity = $entity;
+    }
+
     public function getDefinition(): EntityDefinition
     {
         return new EntityDefinitionMock();
@@ -36,7 +45,7 @@ class EntityRepositoryMock implements EntityRepositoryInterface
 
     public function search(Criteria $criteria, Context $context): EntitySearchResult
     {
-        return new EntitySearchResult(0, new EntityCollection([]), null, $criteria, $context);
+        return new EntitySearchResult(0, new EntityCollection(array_filter([$this->entity])), null, $criteria, $context);
     }
 
     public function update(array $data, Context $context): EntityWrittenContainerEvent