+++ tests/Payone/Webhook/Handler/TransactionStatusWebhookHandlerTest.php
@@ -21,6 +21,8 @@ use Shopware\Core\Checkout\Test\Cart\Common\Generator;
 use Shopware\Core\Defaults;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
+use Shopware\Core\System\Currency\CurrencyEntity;
+use Shopware\Core\System\StateMachine\Aggregation\StateMachineState\StateMachineStateEntity;
 use Shopware\Core\System\StateMachine\StateMachineRegistry;
 use Shopware\Core\System\StateMachine\Transition;
 
@@ -53,19 +55,19 @@ class TransactionStatusWebhookHandlerTest extends TestCase
             $context
         );
 
-        $transactionStatusService = TransactionStatusWebhookHandlerFactory::createTransactionStatusService(
-            $stateMachineRegistry,
-            []
-        );
-
         $orderTransactionEntity = new OrderTransactionEntity();
         $orderTransactionEntity->setId(Constants::ORDER_TRANSACTION_ID);
+
+        $currency = new CurrencyEntity();
+        $currency->setId(Constants::CURRENCY_ID);
+        $currency->setDecimalPrecision(2);
 
         $orderEntity = new OrderEntity();
         $orderEntity->setId(Constants::ORDER_ID);
         $orderEntity->setSalesChannelId(Defaults::SALES_CHANNEL);
         $orderEntity->setAmountTotal(100);
         $orderEntity->setCurrencyId(Constants::CURRENCY_ID);
+        $orderEntity->setCurrency($currency);
 
         $paymentMethodEntity = new PaymentMethodEntity();
         $paymentMethodEntity->setHandlerIdentifier(PayoneCreditCardPaymentHandler::class);
@@ -81,6 +83,16 @@ class TransactionStatusWebhookHandlerTest extends TestCase
         ];
         $orderTransactionEntity->setCustomFields($customFields);
 
+        $stateMachineState = new StateMachineStateEntity();
+        $stateMachineState->setTechnicalName('');
+        $orderTransactionEntity->setStateMachineState($stateMachineState);
+
+        $transactionStatusService = TransactionStatusWebhookHandlerFactory::createTransactionStatusService(
+            $stateMachineRegistry,
+            [],
+            $orderTransactionEntity
+        );
+
         $paymentTransaction = PaymentTransaction::fromOrderTransaction($orderTransactionEntity, $orderEntity);
 
         $transactionData = [
@@ -91,7 +103,7 @@ class TransactionStatusWebhookHandlerTest extends TestCase
 
         $transactionDataHandler = $this->createMock(TransactionDataHandlerInterface::class);
         $transactionDataHandler->expects($this->once())->method('getPaymentTransactionByPayoneTransactionId')->willReturn($paymentTransaction);
-        $transactionDataHandler->expects($this->once())->method('enhanceStatusWebhookData')->willReturn($transactionData);
+        $transactionDataHandler->expects($this->once())->method('getCustomFieldsFromWebhook')->willReturn($transactionData);
 
         $transactionStatusHandler = TransactionStatusWebhookHandlerFactory::createHandler(
             $transactionStatusService,
@@ -121,21 +133,19 @@ class TransactionStatusWebhookHandlerTest extends TestCase
             $context
         );
 
-        $transactionStatusService = TransactionStatusWebhookHandlerFactory::createTransactionStatusService(
-            $stateMachineRegistry,
-            [
-                'paymentStatusAppointed' => 'paid',
-            ]
-        );
-
         $orderTransactionEntity = new OrderTransactionEntity();
         $orderTransactionEntity->setId(Constants::ORDER_TRANSACTION_ID);
+
+        $currency = new CurrencyEntity();
+        $currency->setId(Constants::CURRENCY_ID);
+        $currency->setDecimalPrecision(2);
 
         $orderEntity = new OrderEntity();
         $orderEntity->setId(Constants::ORDER_ID);
         $orderEntity->setSalesChannelId(Defaults::SALES_CHANNEL);
         $orderEntity->setAmountTotal(100);
         $orderEntity->setCurrencyId(Constants::CURRENCY_ID);
+        $orderEntity->setCurrency($currency);
 
         $paymentMethodEntity = new PaymentMethodEntity();
         $paymentMethodEntity->setHandlerIdentifier(PayoneCreditCardPaymentHandler::class);
@@ -151,6 +161,18 @@ class TransactionStatusWebhookHandlerTest extends TestCase
         ];
         $orderTransactionEntity->setCustomFields($customFields);
 
+        $stateMachineState = new StateMachineStateEntity();
+        $stateMachineState->setTechnicalName('');
+        $orderTransactionEntity->setStateMachineState($stateMachineState);
+
+        $transactionStatusService = TransactionStatusWebhookHandlerFactory::createTransactionStatusService(
+            $stateMachineRegistry,
+            [
+                'paymentStatusAppointed' => 'paid',
+            ],
+            $orderTransactionEntity
+        );
+
         $paymentTransaction = PaymentTransaction::fromOrderTransaction($orderTransactionEntity, $orderEntity);
 
         $transactionData = [
@@ -161,7 +183,7 @@ class TransactionStatusWebhookHandlerTest extends TestCase
 
         $transactionDataHandler = $this->createMock(TransactionDataHandlerInterface::class);
         $transactionDataHandler->expects($this->once())->method('getPaymentTransactionByPayoneTransactionId')->willReturn($paymentTransaction);
-        $transactionDataHandler->expects($this->once())->method('enhanceStatusWebhookData')->willReturn($transactionData);
+        $transactionDataHandler->expects($this->once())->method('getCustomFieldsFromWebhook')->willReturn($transactionData);
 
         $transactionStatusHandler = TransactionStatusWebhookHandlerFactory::createHandler(
             $transactionStatusService,