+++ CHANGELOG.md
@@ -37,3 +37,17 @@ Bugfixes
 Known Incompatibilities
 
 * Backurlhandling in Shopware 6.2 is currently broken. If a customer gets redirected to their favorite payment method but decides to cancel and choose another one, no PAYONE payment methods are available. We're working on a fix to enable correct handling of this use case.
+
+# 2.0.0
+
+New Features
+
+* Enables the switch to PAYONE payment methods after ordering
+* Multiple PAYONE transactions can now be handled in the administration per order
+* New payment method: Prepayment
+
+Maintenance
+
+* Added compatibility for new status transitions in Shopware 6.2
+* Shopware 6.2.x support
+* We had to drop support for Shopware <6.2.0